@@ -7,10 +7,10 @@
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.8.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^8.0.4",
-    "react-router-dom": "5.3",
+    "react-router-dom": "^6.4.2",
     "react-scripts": "^5.0.1",
     "web-vitals": "^1.1.1"
   },
